move toasts inside useFarcasterAction
