+++ src/requirements/Farcaster/components/FarcasterActionFollow.tsx
@@ -1,16 +1,11 @@
-import { Text } from "@chakra-ui/react"
 import { RequirementButton } from "components/[guild]/Requirements/components/RequirementButton"
 import { useRequirementContext } from "components/[guild]/Requirements/components/RequirementContext"
 import useUser from "components/[guild]/hooks/useUser"
-import useShowErrorToast from "hooks/useShowErrorToast"
-import useToast from "hooks/useToast"
 import useFarcasterAction from "../hooks/useFarcasterAction"
 import { useFarcasterUser } from "../hooks/useFarcasterUsers"
 
 export default function FarcasterActionFollow() {
   const { farcasterProfiles } = useUser()
-  const toast = useToast()
-  const showErrorToast = useShowErrorToast()
 
   const { data, roleId } = useRequirementContext<
     "FARCASTER_FOLLOW" | "FARCASTER_FOLLOWED_BY"
@@ -18,29 +13,7 @@ export default function FarcasterActionFollow() {
 
   const { data: farcasterUser } = useFarcasterUser(data.id)
 
-  const { onSubmit, isLoading, response } = useFarcasterAction(roleId, "follow", {
-    onSuccess: () => {
-      toast({
-        status: "success",
-        description: (
-          <>
-            <Text>
-              {farcasterUser
-                ? `Successfully followed ${farcasterUser.display_name}`
-                : "Farcaster follow successful"}
-            </Text>
-            <Text fontSize={"sm"}>
-              It might take some time until it shows up in external clients, such as
-              Warpcast
-            </Text>
-          </>
-        ),
-      })
-    },
-    onError: (error) => {
-      showErrorToast(error)
-    },
-  })
+  const { onSubmit, isLoading, response } = useFarcasterAction(roleId, "follow")
 
   if (!farcasterProfiles || !farcasterProfiles?.[0]?.fid || response) {
     return null
@@ -56,7 +29,9 @@ export default function FarcasterActionFollow() {
         onSubmit(data.id.toString())
       }}
     >
-      Follow {farcasterUser?.display_name ?? "on Farcaster"}
+      {farcasterUser?.display_name
+        ? `Follow ${farcasterUser?.display_name}`
+        : "Follow"}
     </RequirementButton>
   )
 }
